% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bweight_centiles}
\alias{bweight_centiles}
\title{Newborn birthweight centile data}
\format{
## `bweight_centiles`
A list with three data frames of 161 rows and 3 columns:
\describe{
  \item{gestational_age_at_birth_weeks}{Gestational age of newborn in weeks}
  \item{C10}{10th percentile of birth weights in kg}
  \item{C90}{90th percentile of birth weights in kg}
}
}
\source{
Provided by Eric Ohuma, London School of Hygiene and Tropical Medicine.
}
\usage{
bweight_centiles
}
\description{
A list with sex-specific tables containing 10th and 90th percentile values for birthweights (in kg) in newborns with
and without known sexes.
}
\keyword{datasets}
