% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bweight_data}
\alias{bweight_data}
\title{Newborn birthweight data}
\format{
## `bweight_data`
A data frame of 90 rows and 4 columns:
\describe{
  \item{birthweight_grams}{Birthweight of newborn in grams}
  \item{gestational_age_at_birth_weeks}{Gestational age of newborn in weeks}
  \item{newborn_sex}{Sex of newborn, if known}
  \item{birthweight_in_kg}{Birthweight of newborn in kilograms}
}
}
\source{
Provided by Eric Ohuma, London School of Hygiene and Tropical Medicine.
}
\usage{
bweight_data
}
\description{
A dataframe containing sex, gestational age and birthweight data for 90 newborns.
}
\keyword{datasets}
