% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bweight_data_annotated}
\alias{bweight_data_annotated}
\title{Newborn birthweight centile data}
\format{
## `bweight_data_annotated`
A list with three data frames of 161 rows and 3 columns:
\describe{
  \item{birthweight_grams}{Birthweight of newborn in grams}
  \item{gestational_age_at_birth_weeks}{Gestational age of newborn in weeks}
  \item{newborn_sex}{Sex of newborn, if known}
  \item{birthweight_in_kg}{Birthweight of newborn in kilograms}
  \item{size_for_ga}{Size for gestational age of newborn (one of SGA, AGA or LGA)}
}
}
\source{
Provided by Eric Ohuma, London School of Hygiene and Tropical Medicine; edited by Simon Parker.
}
\usage{
bweight_data_annotated
}
\description{
A dataframe containing sex, gestational age and birthweight data for 90 newborns, as well as annotations of size for
gestational age for each newborn.
}
\keyword{datasets}
