% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries_sfga.R
\name{get_sfga_summaries}
\alias{get_sfga_summaries}
\title{Generate table of counts of size for gestational age classications}
\usage{
get_sfga_summaries(sfga.vector)
}
\arguments{
\item{sfga.vector}{Vector containing size for gestational age classifications (i.e. 'SGA', 'AGA', 'LGA').}
}
\value{
Data frame with three columns (one for each size for gestational age classifications) and one row with counts for each classification in sfga.vector
}
\description{
Generate table of counts of size for gestational age classications
}
