% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_bweight.R
\name{get_size_for_gestational_age}
\alias{get_size_for_gestational_age}
\title{Get size for gestational age relative to gestational age and sex (if known) in newborns using birthweight in kg.}
\usage{
get_size_for_gestational_age(sex, gestational.age.at.birth, birthweight.in.kg)
}
\arguments{
\item{sex}{The newborn's sex (if known). No default.}

\item{gestational.age.at.birth}{Gestational age of newborn at birth. No default.}

\item{birthweight.in.kg}{Newborn birthweight, measured in kg. No default.}
}
\value{
Size classification for given parameters. Will be one of 'SGA', 'AGA' or 'LGA'.
}
\description{
Get size for gestational age relative to gestational age and sex (if known) in newborns using birthweight in kg.
}
