% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bweight_centiles.R
\name{read_bweight_centiles}
\alias{read_bweight_centiles}
\title{Read birth weight centile data from CSV files with a specific format.}
\usage{
read_bweight_centiles(bweight.centiles.path)
}
\arguments{
\item{bweight.centiles.path}{Path to correctly-formatted CSV with birthweight centile data. No default.}
}
\value{
List of three data frames containing 10th and 90th percentile birthweight values for each gestational age at birth.
}
\description{
Read birth weight centile data from CSV files with a specific format.
}
