% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bweight_data.R
\name{read_bweight_data}
\alias{read_bweight_data}
\title{Read in birth weight CSVs with a specific format}
\usage{
read_bweight_data(bweight.csv.path)
}
\arguments{
\item{bweight.csv.path}{Path to birthweight CSV. No default.}
}
\description{
Reads in CSV containing birthweight data and automatically renames columns to be more R-friendly.
}
